requirements.txt
numpy
openai-whisper
sounddevice
langchain
langchain-community
nltk
torch
transformers
rich